requirements.txt
flask==3.0.3
azure-cosmos==4.7.0
azure-storage-blob==12.23.1
Werkzeug==3.0.3
gunicorn==23.0.0
